+++ src/v3/codec/packet.rs
@@ -1,4 +1,4 @@
-use std::num::NonZeroU16;
+use std::{fmt, num::NonZeroU16};
 
 use bytes::Bytes;
 use bytestring::ByteString;
@@ -84,7 +84,7 @@ impl Connect {
     }
 }
 
-#[derive(Debug, PartialEq, Clone)]
+#[derive(PartialEq, Clone)]
 /// Publish message
 pub struct Publish {
     /// this might be re-delivery of an earlier attempt to send the Packet.
@@ -98,6 +98,19 @@ pub struct Publish {
     pub packet_id: Option<NonZeroU16>,
     /// the Application Message that is being published.
     pub payload: Bytes,
+}
+
+impl fmt::Debug for Publish {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("Publish")
+            .field("packet_id", &self.packet_id)
+            .field("topic", &self.topic)
+            .field("dup", &self.dup)
+            .field("retain", &self.retain)
+            .field("qos", &self.qos)
+            .field("payload", &"<REDACTED>")
+            .finish()
+    }
 }
 
 #[derive(Debug, PartialEq, Copy, Clone)]